feat: asociar platos poa usuario y restringir acceso por autor

+++ backend/src/controllers/dishController.ts
@@ -4,6 +4,10 @@ import { Dish } from "../models/dish";
 //Funcion para crear un nuevo plato 
 export const createDish = async (req: Request, res: Response) => {
     try{
+
+        //Validar la sesion del usuario;
+        const userId = req.user?.id;
+
         //Extrae las variables del cuerpo de la interfaz
         const {image, dishName, description, price} = req.body;
 
@@ -12,7 +16,8 @@ export const createDish = async (req: Request, res: Response) => {
             image,
             dishName,
             description,
-            price
+            price,
+            user: userId, //Asociar los datos al usuario autenticado
         });
 
         //Guarda el nuevo plato en la base de datos.
@@ -28,16 +33,18 @@ export const createDish = async (req: Request, res: Response) => {
 }
 
 //Funcion para crear un nuevo plato 
-export const getAllDishes = async (req: Request, res: Response) => {
+export const getUserDishes = async (req: Request, res: Response) => {
     try{
+
+        const userId = req.user?.id;
         //Busca todos los platos
-        const dishes = await Dish.find();
+        const dishes = await Dish.find({user: userId});
 
         //Envia la respuesta con el estado y la lista de los platos
         res.status(200).json({data: dishes})
     }catch(error){
         //Envia una respuesta del error con el mensaje
-        res.status(500).json({ error: "Error al obtener los datos" });
+        res.status(500).json({ error: "Error al obtener los platos del usuario" });
     }
 }
 
@@ -47,7 +54,20 @@ export const updateDish = async (req: Request, res: Response) => {
     const {id} = req.params;
     //Obtiene los datos del cuerpo 
     const updateData = req.body;
+    const userId = req.user?.id;
     try{
+
+        //Buscar el plato por id
+        const dish = await Dish.findById(id);
+
+        //Si el plato no existe, enviar respuesta.
+         if (!dish) 
+            return res.status(404).json({ message: "Plato no encontrado" })
+        
+         //Validacion para verificar que el usuario de sesion es el mismo que creo el plato
+        if (dish.user.toString() !== userId) {
+        return res.status(403).json({ message: "No tienes permisos para modificar este plato" });
+        }
         
         //Busca el id del plato, pasa los valores y lo actualiza.
         const updateDish =  await Dish.findByIdAndUpdate(
@@ -68,13 +88,26 @@ export const updateDish = async (req: Request, res: Response) => {
 export const removeDish = async (req: Request, res: Response) => {
     //Toma el id de la URL
     const {id} = req.params;
+    const userId = req.user?.id;
     try{
 
-    //Busca el id del plato, pasa los valores y lo actualiza.
-    await Dish.findByIdAndDelete(id);
+         //Buscar el plato por id
+        const dish = await Dish.findById(id);
 
-    //Envia la respuesta con el estado y los nuevos datos del plato
-    return res.status(200).json({ message: "Plato eliminado de la carta"});
+        //Si el plato no existe, enviar respuesta.
+         if (!dish) 
+            return res.status(404).json({ message: "Plato no encontrado" })
+        
+         //Validacion para verificar que el usuario de sesion es el mismo que creo el plato
+        if (dish.user.toString() !== userId) {
+        return res.status(403).json({ message: "No tienes permisos para modificar este plato" });
+        }
+
+        //Busca el id del plato, pasa los valores y lo actualiza.
+        await Dish.findByIdAndDelete(id);
+
+        //Envia la respuesta con el estado y los nuevos datos del plato
+        return res.status(200).json({ message: "Plato eliminado de la carta"});
     }catch(error){
         //Envia una respuesta del error con el mensaje
     return res.status(500).json({ error: "Error al eliminar el plato" });
